+++ mips/Makefile
@@ -1,6 +1,7 @@
 # vim: tabstop=8 shiftwidth=8 noexpandtab:
 
 SOURCES_C = atomic.c \
+	    cbus.c \
 	    clock.c \
 	    context.c \
 	    cpu.c \
